Clean and polish
+ We can eliminate a few implementations that only existed to satisfy the type system if we move their constructors to static factory methods
+ Rename to GroupChangingGatherer and combine tests
+ Combine FilterChangingGatherer tests
